fix(agent): reset _previous_position per episode + docstring types (review)
- reset_environment: clear self._previous_position to None so the first step of a
  new episode computes a zero STAM movement delta instead of a spurious delta against
  the prior episode's last position (pre-existing; surfaced reviewing the float-position
  sensing change).
- env.py: update the `position` parameter docstrings of the three widened gradient
  methods (_compute_food_gradient_vector, _compute_predator_gradient_vector,
  get_separated_gradients) from tuple[int, ...] to tuple[float, ...] to match their
  signatures. Other int-position docstrings belong to unchanged signatures (left as-is).

+++ packages/quantum-nematode/quantumnematode/env/env.py
@@ -2099,7 +2099,7 @@ class DynamicForagingEnvironment(BaseEnvironment):
 
         Parameters
         ----------
-        position : tuple[int, ...]
+        position : tuple[float, ...]
             Position to query gradient at.
 
         Returns
@@ -2147,7 +2147,7 @@ class DynamicForagingEnvironment(BaseEnvironment):
 
         Parameters
         ----------
-        position : tuple[int, ...]
+        position : tuple[float, ...]
             Position to query gradient at.
 
         Returns
@@ -2357,7 +2357,7 @@ class DynamicForagingEnvironment(BaseEnvironment):
 
         Parameters
         ----------
-        position : tuple[int, ...]
+        position : tuple[float, ...]
             Position to query local gradients at.
         disable_log : bool
             Whether to disable debug logging.
